This is an auto push

+++ Proyecto-2019EJB/ejbModule/bean/database/PostgresBeanLocal.java
@@ -55,6 +55,6 @@ public interface PostgresBeanLocal {
 	
 	DtoInfoScooters reporteInfoScooter() throws Exception;
 	float reporteGanancias(Timestamp inicio, Timestamp fin) throws Exception;
-	double reportesCantAlquileres(Timestamp inicio, Timestamp fin) throws Exception;
+	int reportesCantAlquileres(Timestamp inicio, Timestamp fin) throws Exception;
 	List<DtoRakingUsuarios> reporteRakingUsuarios() throws Exception;
 }
